server/requirements.txt
# Python 3.10.14
fastapi==0.112.0
flask==3.0.3
flask_sqlalchemy==3.1.1
langchain-community==0.2.11
langchain==0.2.12
loguru==0.7.2
openai==1.38.0
pymilvus==2.4.4
python-docx==1.1.2
python-dotenv==1.0.1
pytz
qiniu==7.13.2
requests==2.32.3
sentence-transformers==3.0.1
supervisor==4.2.5
uvicorn==0.30.4
jieba==0.42.1
nltk==3.9.1
volcengine==1.0.151